prueba server 6.0

+++ src/controllers/products.controller.js
@@ -1,52 +1,39 @@
 import ProductsServices from "../services/products.service.js";
 import { socketServer } from "../server.js";
 import mongoose from "mongoose";
+import CustomeError from "../services/errors/customeError.js";
+import { productError } from "../services/errors/errorMessages/product.error.js";
 
 class ProductController {
   constructor() {
     this.productService = new ProductsServices();
   };
 
-  async getProducts(req, res) {
-    try {
-      const products = await this.productService.getProducts(req.query);
-      res.send(products);
-    } catch (error) {
-      res
-        .status(500)
-        .send({ status: "error", message: "Error fetching products." });
-      console.log(error);
-    };
-  };
-
-  async getProductById(req, res) {
-    try {
-      const pid = req.params.pid;
-      console.log("Product ID:", pid);
-      const product = await this.productService.getProductById(pid);
-      if (product) {
-        res.json(product);
-        return;
-      } else {
-        res
-          .status(404)
-          .send({ status: "error", message: "Product not found." });
-        return;
-      }
-    } catch (error) {
-      console.error("Error fetching product by id:", error);
-      res
-        .status(500)
-        .send({ status: "error", message: "Error fetching product by id." });
-      return;
-    };
-  };
-
   async createProduct(req, res) {
     let { title, description, code, price, status, stock, category, thumbnail, } = req.body;
-    console.log("Received thumbnail:", thumbnail);
-    if (!title || !description || !code || !price || !status || !stock || !category || !thumbnail) return res.status(400).send({ status: "error", error: "Valores incompletos" });
+    const owner = req.user_id;
+    if (!title) {
+      res.status(400).send({ status: "error", message: "Error no se cargo el campo Titulo" })
+    };
+    if (!description) {
+      res.status(400).send({ status: "error", message: "Error no se cargo el campo Description" })
+    };
+    if (!code) {
+      res.status(400).send({ status: "error", message: "Error no se cargo el campo Codigo" })
+    };
+    if (!price) {
+      res.status(400).send({ status: "error", message: "Error no se cargo el campo Price" })
+    };
     status = !status && true;
+    if (!stock) {
+      res.status(400).send({ status: "error", message: "Error no se cargo el campo Stock" })
+    };
+    if (!category) {
+      res.status(400).send({ status: "error", message: "Error no se cargo el campo Category" })
+    };
+    if (!thumbnail) {
+      res.status(400).send({ status: "error", message: "Error no se cargo el campo Thumbnail" })
+    };
     try {
       const wasAdded = await this.productService.createProduct({
         title,
@@ -56,15 +43,16 @@ class ProductController {
         status,
         stock,
         category,
-        thumbnail: `${req.protocol}://${req.hostname}:${process.env.PORT}/images/${file.filename}`,  // VER ESTO // 
+        thumbnail: `${req.protocol}://${req.hostname}:${process.env.PORT}/images/${file.filename}`,
+        owner,
       });
       if (wasAdded && wasAdded._id) {
-        console.log("Producto añadido correctamente:", wasAdded);
         res.send({
           status: "ok",
           message: "El Producto se agregó correctamente!",
+          productId: wasAdded._id,
         });
-        socketServer.emit("product_created", { _id: wasAdded._id, title, description, code, price, status, stock, category, thumbnail });
+        socketServer.emit("product_created", { _id: wasAdded._id, title, description, code, price, status, stock, category, thumbnail, owner, });
         return;
       } else {
         console.log("Error al añadir producto, wasAdded:", wasAdded);
@@ -80,6 +68,48 @@ class ProductController {
         .status(500)
         .send({ status: "error", message: "Internal server error." });
       return;
+    };
+  };
+
+  async getProducts(req, res) {
+    try {
+      const products = await this.productService.getProducts(req.query);
+      res.send(products);
+    } catch (error) {
+      const productErr = new CustomeError({
+        name: "Product Fetch Error",
+        message: "Error al obtener los productos",
+        code: 500,
+        cause: error.message,
+      });
+      req.logger.error(productErr);
+      res.status(productErr.code).send({ status: "error", message: "Error al obtener los productos" })
+    }
+  }
+
+  async getProductById(req, res) {
+    try {
+      const pid = req.params.pid;
+      if (!mongoose.Types.ObjectId.isValid(pid)) {
+        throw new CustomeError({
+          name: "Invalid ID",
+          message: "El ID no es correcto",
+          code: 400,
+          cause: productError(pid),
+        });
+      };
+      const product = await this.productService.getProductById(pid);
+      if (!product) {
+        throw new CustomeError({
+          name: "Product not found",
+          message: "El producto no pudo ser encontrado",
+          code: 404,
+        });
+      };
+      res.status(200).json({ status: "success", data: product });
+      return;
+    } catch (error) {
+      next(error)
     };
   };
 
@@ -137,6 +167,14 @@ class ProductController {
         });
         return;
       }
+      if (!req.user || (req.user.role !== "admin" && (!product.owner || req.user._id.toString() !== product.owner.toString()))) {
+        console.log("Operación no permitida: el usuario no tiene derechos para eliminar este producto.");
+        res.status(403).send({
+          status: "error",
+          message: "No tiene permiso para eliminar este producto.",
+        });
+        return;
+      };
       const wasDeleted = await this.productService.deleteProduct(pid);
       if (wasDeleted) {
         console.log("Producto eliminado exitosamente");
